Core: Fixed addon loading errors in binary releases
Since addons aren't included in the binary release by default, attempting to load them will fail. In fact, if the WebClient is packaged as-is, there won't even be an addon folder currently, which crashed the startup routines previously.

It is expected that addons will be packaged in a proper release, but in that case they will need to adapt the build scripts that are provided as a guideline anyway.