Refactor: Renamed ElectronAppLoader to JsonConfigLoader
Some are of the opinion JSON is a poor choice for configuration files.

Personally, I don't think it's that bad, since we only use them where they can be automatically generated, loaded, and validated (no comments are needed).

Also, I'd rather invent my own config language than use YAML or any of the other abominations that are suggested instead.

I suppose JS scripts would be an option, but they're far more difficult to serialize and possibly validate, so we're not doing that. Maybe later?