Refactor: Flatten the hierarchy of test suite spec descriptors
This allows adding individual test suites on a per-file basis and aggregating them under a single namespace.

Refactor: Move builtin tests before API tests

If tests for builtins fail, there is little point in going ahead and testing the APIs that rely on them. If anything, it might cause weird errors that are needlessly difficult to troubleshoot.