CI: Fix a rare error with electron-mocha on Linux systems
Looks like sometimes the output can't be routed to the xvfb server and it just fails the CI, but only on Linux (since Windows doesn't need this workaround in the first place).

According to the wise people of StackOverflow, this should prevent that error by letting it choose a free slot automatically if the selected one is unavailable.

Will it work? I have no idea. I guess we'll see if there ever is another failed CI run with this error...