API: Add a BuiltinAudioDecoder to allow using the Resource cache
While this doesn't do anything of significance, it allows the WebAudio API to load audio files from disk using the C_Decoding API, which means they'll be provided as cached resources.

The alternative is to simply use BJS XHR-based loading, but that incurs a lot of overhead since it will load the file from disk every time. There's still overhead for copying buffers and decoding the ArrayBuffer to an AudioBuffer internally, but that'll require a more elaborate solution to eliminate.