Repo: Update addon test script to never skip registered tests
This really isn't great, but it's the only way I've found to ALWAYS execute all tests that were registered after the test runner has been started. Previously it would often skip them, presumably because there was some async loading still in progress.

The trade-off is that only files following the fixed pattern (test-something.js), in a subfolder (of any name, such as test or Tests) will be detected. Not sure if there is a better approach, but I don't currently see one given the requirement that arbitrary tests must be detected if they're registered on demand.