API: Add an experimental TestSuite builtin
I really don't like the design of the testing framework, so this is an attempt at hiding it away and making the test files more readable (albeit not a great one...).

We also want to ensure there's no resource leakage, so I've added some basic checks that likely need to be improved later.