API: Fix decoding of byte-aligned zero-padded BMP files
Some BMP files appear to have empty (zero) bytes at the end, which must be used to byte-align the file in memory. I've only seen this happen when the total file size isn't divisible by four, hence this educated guess.

We can simply ignore it, because it's just padding, so raising an error here is likely uncalled for.